pyproject.toml
[tool.poetry]
name = "geoffrey"
version = "0.1.0"
description = ""
authors = ["Daniel Torres <torresestradadaniel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.1.0"
asyncio = "^3.4.3"
python-dotenv = "^0.21.0"
click = "^8.1.3"
boto3 = "^1.26.34"
python-telegram-bot = "^20.1"
telethon = "^1.27.0"
pyyaml = "^6.0"
rich = "^13.3.2"




[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
